requirements.txt
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy @ file:///croot/contourpy_1732540045555/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
einops==0.8.1
filelock==3.13.1
fonttools @ file:///croot/fonttools_1737039080035/work
fsspec==2024.6.1
fvcore==0.1.5.post20221221
huggingface-hub==0.29.2
idna==3.10
imageio==2.37.0
iopath==0.1.10
Jinja2==3.1.4
kiwisolver @ file:///croot/kiwisolver_1737039087198/work
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.10.0
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpld3 @ file:///croot/mpld3_1696539159604/work
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.3
numpy @ file:///croot/numpy_and_numpy_base_1725470312869/work/dist/numpy-2.0.1-cp310-cp310-linux_x86_64.whl#sha256=120568f3fd675f59e4cb6de79a5c193b4067d90878d8bab2040fda8e3a2df6fa
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
opencv-python==4.11.0.86
packaging @ file:///croot/packaging_1734472117206/work
pandas==2.2.3
pillow==11.0.0
portalocker==3.1.1
prettytable==3.15.1
pyparsing @ file:///croot/pyparsing_1731445506121/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
seaborn==0.13.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sympy==1.13.1
tabulate==0.9.0
termcolor==2.5.0
tifffile==2025.2.18
timm==1.0.15
tokenizers==0.21.0
torch==2.6.0+cu118
torch-dct==0.1.6
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tqdm==4.67.1
transformers==4.49.0
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
unicodedata2 @ file:///croot/unicodedata2_1736541023050/work
urllib3==2.3.0
wcwidth==0.2.13
yacs==0.1.8
